separate all creation of HTML elements in views
